move server to internal
